Fixed Hellfork and Soulfork not dropping cooked food when one-shotting mobs, re-added Tridents item tag
