+++ src/main/java/org/ladysnake/impaled/common/Impaled.java
@@ -3,12 +3,15 @@ package org.ladysnake.impaled.common;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.loot.v3.LootTableEvents;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.item.Item;
 import net.minecraft.loot.LootPool;
 import net.minecraft.loot.LootTables;
 import net.minecraft.loot.condition.LootCondition;
 import net.minecraft.loot.condition.RandomChanceLootCondition;
 import net.minecraft.loot.entry.ItemEntry;
 import net.minecraft.loot.provider.number.UniformLootNumberProvider;
+import net.minecraft.registry.RegistryKeys;
+import net.minecraft.registry.tag.TagKey;
 import net.minecraft.util.Identifier;
 import org.ladysnake.impaled.common.init.ImpaledEntityTypes;
 import org.ladysnake.impaled.common.init.ImpaledItems;
@@ -17,6 +20,7 @@ import org.ladysnake.impaled.common.util.TargetingUtil;
 
 public class Impaled implements ModInitializer {
 	public static final String MOD_ID = "impaled";
+    public static final TagKey<Item> TRIDENTS = TagKey.of(RegistryKeys.ITEM, id("tridents"));
 
 	public static boolean isHoldingSoulfork(LivingEntity living) {
 		return living.getMainHandStack().getItem() == ImpaledItems.SOULFORK || living.getOffHandStack().getItem() == ImpaledItems.SOULFORK;
