+++ src/main/java/org/ladysnake/impaled/mixin/mialeemisc/PlayerEntityMixin.java
@@ -1,0 +1,43 @@
+package org.ladysnake.impaled.mixin.mialeemisc;
+
+import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import org.jetbrains.annotations.Nullable;
+import org.ladysnake.impaled.common.util.IPlayerTargeting;
+import org.ladysnake.impaled.common.util.TargetingUtil;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(PlayerEntity.class)
+public class PlayerEntityMixin implements IPlayerTargeting {
+    @Unique @Nullable LivingEntity lastTarget;
+    @Unique int targetDecayTime;
+
+    @Override
+    public LivingEntity mialeeMisc$getLastTarget() {
+        return this.lastTarget;
+    }
+
+    @Override
+    public void mialeeMisc$setLastTarget(LivingEntity target) {
+        this.lastTarget = target;
+        this.targetDecayTime = 60;
+        if (target != null) {
+            ClientPlayNetworking.send(new TargetingUtil.TargetPayload(target.getId()));
+        }
+    }
+
+    @Inject(method = "tick", at = @At("TAIL"))
+    private void mialeeMisc$decayTarget(CallbackInfo ci) {
+        if (targetDecayTime > 0) {
+            this.targetDecayTime--;
+            if (this.targetDecayTime == 0) {
+                this.mialeeMisc$setLastTarget(null);
+            }
+        }
+    }
+}
