+++ src/main/resources/fabric.mod.json
@@ -38,6 +38,12 @@
 			"net/minecraft/class_8109": [
 				"org/ladysnake/impaled/common/damage/DamageSourcesExt"
 			]
-		}
+		},
+          "modmenu": {
+               "links": {
+                    "modmenu.discord": "https://discord.ladysnake.org",
+                    "modmenu.wiki": "https://ladysnake.org/wiki/impaled"
+               }
+          }
 	}
 }
